Fix [fingers to tap on keyboard] config not loading.

+++ Limelight/Database/DataManager.m
@@ -78,7 +78,7 @@
         settingsToSave.width = [NSNumber numberWithInteger:width];
         settingsToSave.audioConfig = [NSNumber numberWithInteger:audioConfig];
         settingsToSave.onscreenControls = [NSNumber numberWithInteger:onscreenControls];
-        settingsToSave.keyboardToggleFingers = [NSNumber numberWithInteger:keyboardToggleFingers];
+        settingsToSave.keyboardToggleFingers = [NSNumber numberWithInteger:(uint16_t)keyboardToggleFingers];
         settingsToSave.optimizeGames = optimizeGames;
         settingsToSave.multiController = multiController;
         settingsToSave.swapABXYButtons = swapABXYButtons;
